@@ -3,16 +3,17 @@
   "productName": "ComfyUI",
   "repository": "github:comfy-org/electron",
   "copyright": "Copyright © 2024 Comfy Org",
-  "version": "0.3.32",
+  "version": "0.4.0",
   "homepage": "https://comfy.org",
   "description": "The best modular GUI to run AI diffusion models.",
-  "main": ".vite/build/main.js",
+  "main": ".vite/build/main.cjs",
   "packageManager": "yarn@4.5.0",
+  "type": "module",
   "config": {
-    "frontendVersion": "1.5.10",
-    "comfyVersion": "0.3.7",
+    "frontendVersion": "1.6.1",
+    "comfyVersion": "0.3.8",
     "managerCommit": "52b6e4b2eb6481333b29556788f28633e73bd0a9",
-    "uvVersion": "0.5.4"
+    "uvVersion": "0.5.8"
   },
   "scripts": {
     "clean": "rimraf .vite dist out",
@@ -35,20 +36,20 @@
     "make:nvidia": "yarn run make -- --nvidia",
     "notarize": "node debug/notarize.js",
     "package": "yarn run vite:compile && todesktop build --code-sign=false --async",
-    "prepare": "husky",
+    "postinstall": "node .husky/install.mjs",
     "publish": "yarn run vite:compile && todesktop build --async",
     "publish:staging": "yarn run vite:compile && todesktop build --config=./todesktop.staging.json --async",
     "reset-install": "node scripts/resetInstall.js",
     "sign": "node debug/sign.js",
     "start": "node ./scripts/launchdev.js",
     "test:e2e": "npx playwright test",
-    "test:unit": "jest --config jest.config.js",
+    "test:unit": "jest --config jest.config.cjs",
     "test:update-snapshots": "npx playwright test --update-snapshots",
-    "todesktop:afterPack": "./scripts/todesktop/afterPack.js",
-    "todesktop:beforeInstall": "./scripts/todesktop/beforeInstall.js",
-    "typescript": "yarn run tsc",
-    "vite:compile": "vite build --config vite.main.config.ts && vite build --config vite.preload.config.ts",
-    "vite:types": "vite build --config vite.types.config.ts && node scripts/prepareTypes.js",
+    "todesktop:afterPack": "./scripts/todesktop/afterPack.cjs",
+    "todesktop:beforeInstall": "./scripts/todesktop/beforeInstall.cjs",
+    "typescript": "tsc -p tsconfig.build.json",
+    "vite:compile": "yarn run typescript && vite build --config vite.main.config.ts && vite build --config vite.preload.config.ts",
+    "vite:types": "yarn run typescript && vite build --config vite.types.config.ts && node scripts/prepareTypes.js",
     "release:types": "node scripts/releaseTypes.js",
     "update:frontend": "node scripts/updateFrontend.js"
   },
@@ -57,13 +58,15 @@
     "@electron/notarize": "^2.4.0",
     "@electron/rebuild": "^3.7.1",
     "@electron/windows-sign": "^1.1.3",
+    "@eslint/js": "^9.17.0",
     "@playwright/test": "^1.47.2",
     "@sentry/wizard": "^3.30.0",
     "@todesktop/cli": "^1.9.7",
     "@types/adm-zip": "^0.5.5",
     "@types/electron-squirrel-startup": "^1.0.2",
+    "@types/eslint__js": "^8.42.3",
     "@types/jest": "^29.5.14",
-    "@types/node": "^22.5.0",
+    "@types/node": "^22.10.2",
     "@types/tar": "6.1.13",
     "@types/tmp": "^0.2.6",
     "@types/wait-on": "^5.3.4",
@@ -71,8 +74,10 @@
     "@typescript-eslint/parser": "^5.0.0",
     "electron": "31.3.1",
     "electron-builder": "^25.1.8",
-    "eslint": "^8.0.1",
+    "eslint": "^9.17.0",
     "eslint-plugin-import": "^2.25.0",
+    "eslint-plugin-unicorn": "^56.0.1",
+    "globals": "^15.13.0",
     "husky": "^9.1.6",
     "jest": "^29.7.0",
     "lint-staged": "^15.2.10",
@@ -80,8 +85,9 @@
     "rimraf": "^6.0.1",
     "ts-jest": "^29.2.5",
     "ts-node": "^10.0.0",
-    "typescript": "~5.5.4",
-    "vite": "^5.0.12",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.18.1",
+    "vite": "^6.0.3",
     "vite-plugin-dts": "^4.3.0"
   },
   "keywords": [],
