@@ -3,10 +3,10 @@ import { IPC_CHANNELS } from '../constants';
 import log from 'electron-log/main';
 import { ComfyServerConfig } from '../config/comfyServerConfig';
 import type { SystemPaths } from '../preload';
-import fs from 'fs';
+import fs from 'node:fs';
 import si from 'systeminformation';
 import { ComfyConfigManager } from '../config/comfyConfigManager';
-import path from 'path';
+import path from 'node:path';
 
 export class PathHandlers {
   static readonly REQUIRED_SPACE = 10 * 1024 * 1024 * 1024; // 10GB in bytes
@@ -57,7 +57,7 @@ export class PathHandlers {
           // Check if path is writable
           try {
             fs.accessSync(inputPath, fs.constants.W_OK);
-          } catch (err) {
+          } catch {
             return { isValid: false, error: 'Path is not writable' };
           }
 
@@ -76,7 +76,7 @@ export class PathHandlers {
           log.error('Error validating install path:', error);
           return {
             isValid: false,
-            error: 'Failed to validate install path: ' + error,
+            error: `Failed to validate install path: ${error}`,
           };
         }
       }
