@@ -1,6 +1,6 @@
-import path from 'path';
+import path from 'node:path';
 import log from 'electron-log/main';
-import fs from 'fs';
+import fs from 'node:fs';
 import { InstallOptions } from '../preload';
 import { DEFAULT_SETTINGS } from '../config/comfySettings';
 import { ComfyServerConfig, ModelPaths } from '../config/comfyServerConfig';
@@ -67,6 +67,12 @@ export class InstallWizard {
       'Comfy-Desktop.AutoUpdate': this.installOptions.autoUpdate,
       'Comfy-Desktop.SendStatistics': this.installOptions.allowMetrics,
     };
+
+    if (this.installOptions.device === 'cpu') {
+      settings['Comfy.Server.LaunchArgs'] ??= {};
+      settings['Comfy.Server.LaunchArgs']['cpu'] = '';
+    }
+
     const settingsJson = JSON.stringify(settings, null, 2);
     fs.writeFileSync(settingsPath, settingsJson);
     log.info(`Wrote settings to ${settingsPath}: ${settingsJson}`);
