@@ -1,5 +1,5 @@
 import { app } from 'electron';
-import path from 'path';
+import path from 'node:path';
 
 export function getAppResourcesPath(): string {
   return app.isPackaged ? process.resourcesPath : path.join(app.getAppPath(), 'assets');
