@@ -1,4 +1,4 @@
-import { app, dialog, ipcMain, type Point } from 'electron';
+import { app, dialog, ipcMain } from 'electron';
 import log from 'electron-log/main';
 import * as Sentry from '@sentry/electron/main';
 import { graphics } from 'systeminformation';
@@ -8,15 +8,15 @@ import { ComfySettings } from '../config/comfySettings';
 import { AppWindow } from './appWindow';
 import { ComfyServer } from './comfyServer';
 import { ComfyServerConfig } from '../config/comfyServerConfig';
-import fs from 'fs';
-import { InstallOptions, type ElectronContextMenuOptions } from '../preload';
-import path from 'path';
-import { getModelsDirectory, validateHardware } from '../utils';
+import fs from 'node:fs';
+import { InstallOptions, type ElectronContextMenuOptions, type TorchDeviceType } from '../preload';
+import path from 'node:path';
+import { ansiCodes, getModelsDirectory, validateHardware } from '../utils';
 import { DownloadManager } from '../models/DownloadManager';
 import { ProcessCallbacks, VirtualEnvironment } from '../virtualEnvironment';
 import { InstallWizard } from '../install/installWizard';
 import { Terminal } from '../shell/terminal';
-import { DesktopConfig } from '../store/desktopConfig';
+import { useDesktopConfig } from '../store/desktopConfig';
 import { InstallationValidator } from '../install/installationValidator';
 import { restoreCustomNodes } from '../services/backup';
 import { CmCli } from '../services/cmCli';
@@ -38,7 +38,7 @@ export class ComfyDesktopApp {
   }
 
   public async initialize(): Promise<void> {
-    this.comfySettings.loadSettings();
+    await this.comfySettings.loadSettings();
     this.registerIPCHandlers();
     this.initializeTodesktop();
     await this.setupGPUContext();
@@ -67,7 +67,7 @@ export class ComfyDesktopApp {
       this.terminal?.resize(cols, rows);
     });
 
-    ipcMain.handle(IPC_CHANNELS.TERMINAL_RESTORE, (_event) => {
+    ipcMain.handle(IPC_CHANNELS.TERMINAL_RESTORE, () => {
       return this.terminal?.restore();
     });
   }
@@ -86,11 +86,11 @@ export class ComfyDesktopApp {
       }));
 
       // Combine all GPU info into a single object
-      const allGpuInfo = Object.assign({}, ...gpuInfo);
+      const allGpuInfo = { ...gpuInfo };
       // Set Sentry context with all GPU information
       Sentry.setContext('gpus', allGpuInfo);
-    } catch (e) {
-      log.error('Error getting GPU info: ', e);
+    } catch (error) {
+      log.error('Error getting GPU info: ', error);
     }
   }
 
@@ -123,19 +123,34 @@ export class ComfyDesktopApp {
       log.info('Reinstalling...');
       this.reinstall();
     });
-    ipcMain.handle(IPC_CHANNELS.SEND_ERROR_TO_SENTRY, async (_event, { error, extras }): Promise<string | null> => {
-      try {
-        return Sentry.captureMessage(error, {
-          level: 'error',
-          extra: { ...extras, comfyUIExecutionError: true },
-          tags: {
-            comfyorigin: 'core',
-          },
-        });
-      } catch (err) {
-        log.error('Failed to send error to Sentry:', err);
-        return null;
+    type SentryErrorDetail = {
+      error: string;
+      extras?: Record<string, unknown>;
+    };
+
+    ipcMain.handle(
+      IPC_CHANNELS.SEND_ERROR_TO_SENTRY,
+      async (_event, { error, extras }: SentryErrorDetail): Promise<string | null> => {
+        try {
+          return Sentry.captureMessage(error, {
+            level: 'error',
+            extra: { ...extras, comfyUIExecutionError: true },
+            tags: {
+              comfyorigin: 'core',
+            },
+          });
+        } catch (error_) {
+          log.error('Failed to send error to Sentry:', error_);
+          return null;
+        }
       }
+    );
+    // Restart core
+    ipcMain.handle(IPC_CHANNELS.RESTART_CORE, async (): Promise<boolean> => {
+      if (!this.comfyServer) return false;
+      await this.comfyServer?.kill();
+      await this.comfyServer.start();
+      return true;
     });
   }
 
@@ -143,7 +158,12 @@ export class ComfyDesktopApp {
    * Install ComfyUI and return the base path.
    */
   static async install(appWindow: AppWindow): Promise<string> {
+    const config = useDesktopConfig();
+    if (!config.get('installState')) config.set('installState', 'started');
+
     const validation = await validateHardware();
+    if (typeof validation?.gpu === 'string') config.set('detectedGpu', validation.gpu);
+
     if (!validation.isValid) {
       await appWindow.loadRenderer('not-supported');
       log.error(validation.error);
@@ -151,67 +171,76 @@ export class ComfyDesktopApp {
       await appWindow.loadRenderer('welcome');
     }
 
-    return new Promise<string>((resolve) => {
-      ipcMain.on(IPC_CHANNELS.INSTALL_COMFYUI, async (event, installOptions: InstallOptions) => {
+    return new Promise<string>((resolve, reject) => {
+      ipcMain.on(IPC_CHANNELS.INSTALL_COMFYUI, (_event, installOptions: InstallOptions) => {
         const installWizard = new InstallWizard(installOptions);
-        const { store } = DesktopConfig;
-        store.set('basePath', installWizard.basePath);
+        useDesktopConfig().set('basePath', installWizard.basePath);
 
-        await installWizard.install();
-        store.set('installState', 'installed');
-        if (installWizard.shouldMigrateCustomNodes) {
-          store.set('migrateCustomNodes', installWizard.migrationSource);
+        const { device } = installOptions;
+        if (device !== undefined) {
+          useDesktopConfig().set('selectedDevice', device);
         }
-        appWindow.maximize();
-        resolve(installWizard.basePath);
+
+        installWizard
+          .install()
+          .then(() => {
+            useDesktopConfig().set('installState', 'installed');
+            appWindow.maximize();
+            if (installWizard.shouldMigrateCustomNodes) {
+              store.set('migrateCustomNodesFrom', installWizard.migrationSource);
+            }
+            resolve(installWizard.basePath);
+          })
+          .catch(reject);
       });
     });
   }
 
   async startComfyServer(serverArgs: ServerArgs) {
-    app.on('before-quit', async () => {
+    app.on('before-quit', () => {
       if (!this.comfyServer) {
         return;
       }
 
-      try {
-        log.info('Before-quit: Killing Python server');
-        await this.comfyServer.kill();
-      } catch (error) {
+      log.info('Before-quit: Killing Python server');
+      this.comfyServer.kill().catch((error) => {
         log.error('Python server did not exit properly');
         log.error(error);
-      }
+      });
     });
     log.info('Server start');
-    this.appWindow.loadRenderer('server-start');
+    await this.appWindow.loadRenderer('server-start');
 
-    DownloadManager.getInstance(this.appWindow!, getModelsDirectory(this.basePath));
+    DownloadManager.getInstance(this.appWindow, getModelsDirectory(this.basePath));
 
     this.appWindow.sendServerStartProgress(ProgressStatus.PYTHON_SETUP);
-    const virtualEnvironment = new VirtualEnvironment(this.basePath);
+
+    const config = useDesktopConfig();
+    const selectedDevice = config.get('selectedDevice');
+    const virtualEnvironment = new VirtualEnvironment(this.basePath, selectedDevice);
 
     const processCallbacks: ProcessCallbacks = {
       onStdout: (data) => {
-        log.info(data);
+        log.info(data.replaceAll(ansiCodes, ''));
         this.appWindow.send(IPC_CHANNELS.LOG_MESSAGE, data);
       },
       onStderr: (data) => {
-        log.error(data);
+        log.error(data.replaceAll(ansiCodes, ''));
         this.appWindow.send(IPC_CHANNELS.LOG_MESSAGE, data);
       },
-    };
+    }
+
     await virtualEnvironment.create(processCallbacks);
 
-    const { store } = DesktopConfig;
     const customNodeMigrationError = await this.migrateCustomNodes(store, virtualEnvironment, processCallbacks);
 
-    if (!store.get('Comfy-Desktop.RestoredCustomNodes', false)) {
+    if (!config.get('Comfy-Desktop.RestoredCustomNodes', false)) {
       try {
         await restoreCustomNodes(virtualEnvironment, this.appWindow);
-        store.set('Comfy-Desktop.RestoredCustomNodes', true);
+        config.set('Comfy-Desktop.RestoredCustomNodes', true);
       } catch (error) {
         log.error('Failed to restore custom nodes:', error);
-        store.set('Comfy-Desktop.RestoredCustomNodes', false);
+        config.set('Comfy-Desktop.RestoredCustomNodes', false);
       }
     }
 
@@ -236,7 +265,7 @@ export class ComfyDesktopApp {
     virtualEnvironment: VirtualEnvironment,
     callbacks: ProcessCallbacks
   ) {
-    const customNodeMigrationPath = store.get('migrateCustomNodes');
+    const customNodeMigrationPath = store.get('migrateCustomNodesFrom');
     let customNodeMigrationError: string | null = null;
     if (customNodeMigrationPath) {
       log.info('Migrating custom nodes from: ', customNodeMigrationPath);
@@ -249,16 +278,15 @@ export class ComfyDesktopApp {
           error instanceof Error ? error.message : typeof error === 'string' ? error : 'Error migrating custom nodes.';
       } finally {
         // Always remove the flag so the user doesnt get stuck here
-        store.delete('migrateCustomNodes');
+        store.delete('migrateCustomNodesFrom');
       }
     }
     return customNodeMigrationError;
   }
 
   static async create(appWindow: AppWindow): Promise<ComfyDesktopApp> {
-    const { store } = DesktopConfig;
     // Migrate settings from old version if required
-    const installState = store.get('installState') ?? (await ComfyDesktopApp.migrateInstallState());
+    const installState = useDesktopConfig().get('installState') ?? (await ComfyDesktopApp.migrateInstallState());
 
     // Fresh install
     const loadedPath = installState === undefined ? undefined : await ComfyDesktopApp.loadBasePath();
@@ -279,10 +307,10 @@ export class ComfyDesktopApp {
     const basePath = await ComfyDesktopApp.loadBasePath();
 
     // Migrate config
-    const { store } = DesktopConfig;
+    const config = useDesktopConfig();
     const upgraded = 'upgraded';
-    store.set('installState', upgraded);
-    store.set('basePath', basePath);
+    config.set('installState', upgraded);
+    config.set('basePath', basePath);
     return upgraded;
   }
 
@@ -303,9 +331,8 @@ export class ComfyDesktopApp {
         return null;
       case 'notFound':
         return null;
-      case 'error':
       default:
-        // Explain and quit
+        // 'error': Explain and quit
         // TODO: Support link?  Something?
         await InstallationValidator.showInvalidFileAndQuit(ComfyServerConfig.configPath, {
           message: `Unable to read the YAML configuration file.  Please ensure this file is available and can be read:
@@ -317,7 +344,7 @@ If this problem persists, back up and delete the config file, then restart the a
           defaultId: 0,
           cancelId: 1,
         });
-        throw new Error(/* Unreachable. */);
+        throw new Error('Unreachable');
     }
   }
 
