@@ -1,14 +1,14 @@
 import { app } from 'electron';
 import { IPC_CHANNELS, ServerArgs } from '../constants';
 import { VirtualEnvironment } from '../virtualEnvironment';
-import { rotateLogFiles } from '../utils';
+import { ansiCodes, rotateLogFiles } from '../utils';
 import { getAppResourcesPath } from '../install/resourcePaths';
 import log from 'electron-log/main';
-import path from 'path';
+import path from 'node:path';
 import { ComfyServerConfig } from '../config/comfyServerConfig';
 import { AppWindow } from './appWindow';
 import waitOn from 'wait-on';
-import { ChildProcess } from 'child_process';
+import { ChildProcess } from 'node:child_process';
 
 export class ComfyServer {
   /**
@@ -80,13 +80,13 @@ export class ComfyServer {
   }
 
   static buildLaunchArgs(mainScriptPath: string, args: Record<string, string>) {
-    return [mainScriptPath].concat(
-      Object.entries(args)
-        .map(([key, value]) => [`--${key}`, value])
-        .flat()
+    return [
+      mainScriptPath,
+      ...Object.entries(args)
+        .flatMap(([key, value]) => [`--${key}`, value])
         // Empty string values are ignored. e.g. { cpu: '' } => '--cpu'
-        .filter((value: string) => value !== '')
-    );
+        .filter((value: string) => value !== ''),
+    ];
   }
 
   get launchArgs() {
@@ -98,9 +98,15 @@ export class ComfyServer {
 
   async start() {
     await rotateLogFiles(app.getPath('logs'), 'comfyui', 50);
-    return new Promise<void>(async (resolve, reject) => {
+    return new Promise<void>((resolve, reject) => {
       const comfyUILog = log.create({ logId: 'comfyui' });
       comfyUILog.transports.file.fileName = 'comfyui.log';
+
+      // TODO: Check if electron-log has updated types
+      comfyUILog.transports.file.transforms.push(({ data }) => {
+        // @ts-expect-error electron-log types are broken.  data and return type are `string`.
+        return typeof data === 'string' ? data.replaceAll(ansiCodes, '') : data;
+      });
 
       const comfyServerProcess = this.virtualEnvironment.runPythonCommand(this.launchArgs, {
         onStdout: (data) => {
@@ -130,19 +136,19 @@ export class ComfyServer {
 
       this.comfyServerProcess = comfyServerProcess;
 
-      try {
-        await waitOn({
-          resources: [`${this.baseUrl}/queue`],
-          timeout: ComfyServer.MAX_FAIL_WAIT,
-          interval: ComfyServer.CHECK_INTERVAL,
+      waitOn({
+        resources: [`${this.baseUrl}/queue`],
+        timeout: ComfyServer.MAX_FAIL_WAIT,
+        interval: ComfyServer.CHECK_INTERVAL,
+      })
+        .then(() => {
+          log.info('Python server is ready');
+          resolve();
+        })
+        .catch((error) => {
+          log.error('Server failed to start:', error);
+          reject(new Error('Python Server Failed To Start Within Timeout.'));
         });
-
-        log.info('Python server is ready');
-        resolve();
-      } catch (error) {
-        log.error('Server failed to start:', error);
-        reject('Python Server Failed To Start Within Timeout.');
-      }
     });
   }
 
@@ -158,7 +164,7 @@ export class ComfyServer {
       // Set up a timeout in case the process doesn't exit
       const timeout = setTimeout(() => {
         reject(new Error('Timeout: Python server did not exit within 10 seconds'));
-      }, 10000);
+      }, 10_000);
 
       // Listen for the 'exit' event
       this.comfyServerProcess.once('exit', (code, signal) => {
